requirements.txt
Flask==2.3.3 
gunicorn==23.0.0 
langchain==0.3.24 
langchain-community==0.3.22 
langchain-core==0.3.55 
langchain-pinecone==0.2.5 
pinecone==6.0.2 
huggingface-hub==0.30.2 
sentence-transformers==2.7.0 
transformers==4.39.3 
torch==2.7.0 
python-dotenv==1.1.0 
deep-translator==1.11.4 
langid==1.1.6 
pypdf==5.4.0 
rank-bm25==0.2.2 
requests==2.31.0 
numpy==2.2.5
Django==4.2.7 
django-cors-headers==4.3.1 
psycopg2-binary==2.9.9
cachetools==5.5.2